chore: refactor more out of rfc.ts
