+++ client/utilities/rfc.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment nuxt
 import { vi, test, expect, describe, beforeEach, afterEach } from 'vitest'
 import { DateTime } from 'luxon'
-import { parseRFCId, refsRefRfcIdTxt } from './rfc'
+import { parseRFCId } from './rfc'
 import { NONBREAKING_SPACE } from './strings'
 import rfcRefs from './rfc-refs.json'
 import rfcJsons from './rfc-jsons.json'
@@ -12,6 +12,7 @@ import {
   formatDatePublished,
   parseRfcJsonPubDateToISO
 } from './rfc-converters-utils'
+import { refsRefRfcIdTxt } from './refs'
 import type { ApiClient, Rfc, RfcMetadata } from '~/generated/red-client'
 
 test('parseRFCId', () => {
